update scheduler, use config table

+++ src/main/java/com/example/springboot/service/FileStorageService.java
@@ -2,7 +2,9 @@ package com.example.springboot.service;
 
 import com.example.springboot.constant.Constant;
 import com.example.springboot.models.ESR_inbound_filter_model;
+import com.example.springboot.models.esr_inbound_filter_config_model;
 import com.example.springboot.repositories.ESR_inbound_filter_model_repository;
+import com.example.springboot.repositories.esr_inbound_filter_config_model_repository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.w3c.dom.Document;
@@ -13,6 +15,7 @@ import org.xml.sax.SAXException;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -30,9 +33,31 @@ public class FileStorageService {
     private final Path fileStorageLocation;
     @Autowired
     private final ESR_inbound_filter_model_repository esr_inbound_filter_model_repository;
+    @Autowired
+    private final esr_inbound_filter_config_model_repository esr_inbound_filter_config_model_repository;
     public FileStorageService(ESR_inbound_filter_model_repository esr_inbound_filter_model_repository) {
         this.esr_inbound_filter_model_repository = esr_inbound_filter_model_repository;
         fileStorageLocation = null;
+        esr_inbound_filter_config_model_repository = null;
+    }
+
+    private void routeToVEMS(String uuid) {
+        // Implement routing logic to VEMS
+        ESR_inbound_filter_model existingEntity = esr_inbound_filter_model_repository.findByUuid(uuid);
+        existingEntity.sent = "VEMS";
+        esr_inbound_filter_model_repository.save(existingEntity);
+        System.out.println("Routing XML to VEMS");
+    }
+
+    private void routeToOMS(String uuid) {
+        // Implement routing logic to OMS
+        ESR_inbound_filter_model existingEntity = esr_inbound_filter_model_repository.findByUuid(uuid);
+        existingEntity.sent = "OMS";
+        esr_inbound_filter_model_repository.save(existingEntity);
+        System.out.println("Routing XML to OMS");
+    }
+    public void schdule5mins(String folderPath, String specificPath){
+
     }
     public void checkXmlfile5min(String folderPath, String specificPath){
         Path path = Paths.get(folderPath);
@@ -73,10 +98,72 @@ public class FileStorageService {
                     }
                 }
                 if(!eventType.equals("")&&!uuid.equals("")){
-                    ESR_inbound_filter_model esr_inbound_filter_model = new ESR_inbound_filter_model(eventType,uuid,"a","NO","b",new Date(),new Date());
-                    esr_inbound_filter_model_repository.save(esr_inbound_filter_model);
-                    Files.move(xmlFile, destinationPath, StandardCopyOption.REPLACE_EXISTING);
-                    System.out.println("File move: "+xmlFile.getFileName()+"   "+destinationPath.toString());
+                    String sent_to_system = "";
+                    ESR_inbound_filter_model existingEntity = esr_inbound_filter_model_repository.findByUuid(uuid);
+                    if (existingEntity==null) {
+                        List<esr_inbound_filter_config_model> stateInDB;
+                        stateInDB = esr_inbound_filter_config_model_repository.findAll();
+                        String state = "";
+                        String zipcode = "0";
+                        String contractId = "";
+                        NodeList list = document.getElementsByTagName(eventType);
+                        for(int i=0;i<list.getLength();i++) {
+                            Node node = list.item(i);
+                            if (node.getNodeType() == Node.ELEMENT_NODE) {
+                                Element element = (Element) list.item(0);
+                                contractId =  element.getAttribute("contractId");
+                            }
+                        }
+                        System.out.println("ContractID: "+ contractId);
+                        NodeList addressList = document.getElementsByTagName("PreferredGeoAddress");
+                        for(int i=0;i<addressList.getLength();i++){
+                            Node node = addressList.item(i);
+                            if (node.getNodeType() == Node.ELEMENT_NODE){
+                                Element element = (Element) node;
+                                state = element.getElementsByTagName("State").item(0).getTextContent();
+                                zipcode = element.getElementsByTagName("ZipOrPostalCode").item(0).getTextContent();
+                            }
+                        }
+                        if(state == null && zipcode == null){
+                            String fallbackState = null, fallbackZip = null;
+                            NodeList AddressList1 = document.getElementsByTagName("Address");
+                            for(int i=0;i<AddressList1.getLength();i++){
+                                Node node = AddressList1.item(i);
+                                if (node.getNodeType() == Node.ELEMENT_NODE){
+                                    Element element = (Element) node;
+                                    fallbackState = element.getElementsByTagName("State").item(0).getTextContent();
+                                    fallbackZip = element.getElementsByTagName("ZipOrPostalCode").item(0).getTextContent();
+                                }
+                            }
+                            boolean fallbackStateIsCA = stateInDB.get(0).config_type.equalsIgnoreCase(fallbackState) || "California".equalsIgnoreCase(fallbackState);
+                            boolean fallbackZipIsInCA = (Integer. parseInt(fallbackZip)>90001&&Integer. parseInt(fallbackZip)<96162);
+                            if ((fallbackStateIsCA || fallbackZipIsInCA) && contractId.contains("_R4_")) {
+                                sent_to_system = "VEMS";
+                            } else {
+                                sent_to_system = "OMS";
+                            }
+                        }else {
+                            boolean stateIsCA = state.equalsIgnoreCase(stateInDB.get(0).config_type) || state.equalsIgnoreCase("California");
+                            boolean zipIsInCA = (Integer. parseInt(zipcode)>90001&&Integer. parseInt(zipcode)<96162);
+                            boolean contractIdContainsR4 = contractId.contains("_R4_");
+                            if((stateIsCA || zipIsInCA) && contractIdContainsR4){
+                                sent_to_system = "VEMS";
+                            }else{
+                                sent_to_system = "OMS";
+                            }
+                        }
+                        ESR_inbound_filter_model esr_inbound_filter_model = new ESR_inbound_filter_model(eventType,uuid,"READY",sent_to_system,"",new Date(),new Date());
+                        esr_inbound_filter_model_repository.save(esr_inbound_filter_model);
+                    }else{
+                        if(existingEntity.esr_status.equalsIgnoreCase(":w")){
+                            try {
+                                Files.move(xmlFile, destinationPath, StandardCopyOption.REPLACE_EXISTING);
+                            } catch (IOException e) {
+                                throw new RuntimeException(e);
+                            }
+                            System.out.println("File move: "+xmlFile.getFileName()+"   "+destinationPath.toString());
+                        }
+                    }
                 }
             } catch (ParserConfigurationException e) {
                 throw new RuntimeException(e);
@@ -87,6 +174,8 @@ public class FileStorageService {
             }
 
         });
+
+
 
     }
     public void checkXmlfile10min(){
